Changed int to long long in Fenwick

+++ root/data_structures/FenwickAddSetPointGetSegSum.cpp
@@ -16,37 +16,37 @@ using ll = long long;
 
 class FenwickTree {
 private:
-    int sz;
-    vector<int> tree;
+    ll sz;
+    vector<ll> tree;
 public:
-    explicit FenwickTree(int sz) : sz(sz) {
+    explicit FenwickTree(ll sz) : sz(sz) {
         tree.resize(sz);
     }
 
-    explicit FenwickTree(const vector<int>& src) : sz(int(src.size())) {
+    explicit FenwickTree(const vector<ll>& src) : sz(ll(src.size())) {
         tree.resize(sz);
-        for (int i = 0; i < sz; ++i) {
+        for (ll i = 0; i < sz; ++i) {
             add(i, src[i]);
         }
     }
 
-    void add(int pos, int val) {
+    void add(ll pos, ll val) {
         while (pos < sz) {
             tree[pos] += val;
             pos |= (pos + 1);
         }
     }
 
-    void set(int pos, int val) {
+    void set(ll pos, ll val) {
         add(pos, val - getSum(val));
     }
 
-    int getSum(int l, int r) {
+    ll getSum(ll l, ll r) {
         return getSum(r) - getSum(l - 1);
     }
 
-    int getSum(int r) {
-        int sum = 0;
+    ll getSum(ll r) {
+        ll sum = 0;
         while (r >= 0) {
             sum += tree[r];
             r = (r & (r + 1)) - 1;
